requirements.txt
django==1.11
djangorestframework
idna
